requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
opencv-python>=4.8.1.78
pillow>=10.1.0
numpy>=1.26.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
slack-sdk>=3.23.0
twilio>=8.10.0
sendgrid>=6.10.0
airtable-python-wrapper>=0.15.3
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
pandas>=2.1.3
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
dash>=2.14.2
dash-bootstrap-components>=1.5.0
